requirements.txt
# Core Libraries (Required)
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.8.0

# Machine Learning (Required)
scikit-learn>=1.0.0
joblib>=1.2.0

# Configuration & Serialization (Required)
pyyaml>=5.4.0

# Security & Authentication (Required)
pyjwt>=2.0.0
cryptography>=3.0.0
bcrypt>=3.2.0

# Testing Framework (Required for development)
pytest>=6.0.0

# Optional: TensorFlow for Neural Networks
tensorflow>=2.10.0; extra == "tensorflow"

# Optional: GPU Acceleration (Linux only)
cudf>=24.0.0; sys_platform == "linux" and extra == "gpu"
cuml>=24.0.0; sys_platform == "linux" and extra == "gpu"
cupy-cuda11x>=12.0.0; sys_platform == "linux" and extra == "gpu"

# Optional: Additional Data Processing
pyarrow>=10.0.0; extra == "data"
h5py>=3.0.0; extra == "data"

# Optional: API Framework
fastapi>=0.100.0; extra == "api"
uvicorn>=0.20.0; extra == "api"
pydantic>=2.0.0; extra == "api"

# Optional: Additional Testing
pytest-cov>=4.0.0; extra == "test"
pytest-mock>=3.0.0; extra == "test"

# Optional: Monitoring & Logging
loguru>=0.6.0; extra == "monitoring"
psutil>=5.8.0; extra == "monitoring"
prometheus_client>=0.15.0; extra == "monitoring"

# Optional: Development Tools
black>=22.0.0; extra == "dev"
flake8>=5.0.0; extra == "dev"
mypy>=1.0.0; extra == "dev"

# Optional: Visualization
matplotlib>=3.5.0; extra == "viz"
seaborn>=0.11.0; extra == "viz"
plotly>=5.0.0; extra == "viz"

# Optional: Utilities
tqdm>=4.60.0; extra == "utils"
click>=8.0.0; extra == "utils"
python-dotenv>=0.19.0; extra == "utils"

# Optional: Experiment Tracking
mlflow>=2.0.0; extra == "mlflow"